requirements.txt
requests==2.9.1
sleekxmpp==1.3.1
slackclient==0.16
python-dateutil==1.5